zlib: fix conversion warnings

+++ libs/zlib/src/deflate.c
@@ -1794,7 +1794,7 @@ static block_state deflate_stored(deflate_state *s, int flush)
         return block_done;
 
     /* Fill the window with any remaining input. */
-    have = s->window_size - s->strstart - 1;
+    have = (uint32_t)s->window_size - s->strstart - 1;
     if (s->strm->avail_in > have && s->block_start >= (int64_t)s->w_size) {
         /* Slide the window down. */
         s->block_start -= s->w_size;
